split messages module
