+++ raiden/routing.py
@@ -7,7 +7,7 @@ import structlog
 from eth_utils import to_canonical_address, to_checksum_address
 
 from raiden.exceptions import ServiceRequestFailed
-from raiden.messages import RouteMetadata
+from raiden.messages.metadata import RouteMetadata
 from raiden.network.pathfinding import PFSConfig, query_paths
 from raiden.transfer import channel, views
 from raiden.transfer.state import ChainState, ChannelState, RouteState
