+++ raiden/tests/unit/transfer/mediated_transfer/test_mediatorstate_regression.py
@@ -3,7 +3,8 @@ import random
 
 from eth_utils import keccak
 
-from raiden.messages import message_from_sendevent
+from raiden.messages.decode import balanceproof_from_envelope
+from raiden.messages.encode import message_from_sendevent
 from raiden.tests.utils import factories
 from raiden.tests.utils.events import search_for_item
 from raiden.tests.utils.factories import (
@@ -34,11 +35,7 @@ from raiden.transfer.mediated_transfer.state_change import (
     ReceiveSecretReveal,
     ReceiveTransferRefund,
 )
-from raiden.transfer.state import (
-    NODE_NETWORK_UNREACHABLE,
-    balanceproof_from_envelope,
-    message_identifier_from_prng,
-)
+from raiden.transfer.state import NODE_NETWORK_UNREACHABLE, message_identifier_from_prng
 from raiden.transfer.state_change import Block, ContractReceiveSecretReveal
 from raiden.utils.signer import LocalSigner
