+++ raiden/tests/utils/transfer.py
@@ -10,7 +10,10 @@ from gevent.timeout import Timeout
 from raiden.app import App
 from raiden.constants import EMPTY_SIGNATURE, UINT64_MAX
 from raiden.message_handler import MessageHandler
-from raiden.messages import LockedTransfer, LockExpired, Message, Metadata, RouteMetadata, Unlock
+from raiden.messages.abstract import Message
+from raiden.messages.decode import balanceproof_from_envelope
+from raiden.messages.metadata import Metadata, RouteMetadata
+from raiden.messages.transfers import LockedTransfer, LockExpired, Unlock
 from raiden.tests.utils.factories import make_address, make_secret
 from raiden.tests.utils.protocol import WaitForMessage
 from raiden.transfer import channel, views
@@ -24,7 +27,6 @@ from raiden.transfer.state import (
     HashTimeLockState,
     NettingChannelState,
     PendingLocksState,
-    balanceproof_from_envelope,
     make_empty_pending_locks_state,
 )
 from raiden.utils import random_secret
